New configuration durationOnMaxAgeZero
If a server sends "Cache-Control: max-age=0" this might be a
misconfiguration on the server; in this case we should ignore it (the
behaviour up to now).

It might also mean, that the content potentially already expired; for
this case this commit provides a new configuration parameter so you
reinterpret the duration how long this downloaded file should stay
valid.